+++ CHANGELOG.md
@@ -6,10 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.3.0] - 2020-01-19
 ### Added
 #### Scrobble timeline
 * A dataset that maps artists to their main genres.
-* A script that combines this dataset into a single file.
+* A script that converts this dataset to a single `.json` file.
+
+### Changed
+#### Scrobble timeline
+* Artist names are corrected (duplicates are merged).
 
 ## [1.2.0] - 2019-11-09
 ### Added